financial_ai.py
import pandas as pd
import numpy as np
from prophet import Prophet
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler
import openai
from datetime import datetime, timedelta
import os
from dotenv import load_dotenv

class FinancialAI:
    def __init__(self):
        load_dotenv()
        self.openai_api_key = os.getenv('OPENAI_API_KEY')
        if self.openai_api_key:
            openai.api_key = self.openai_api_key
        self._initialize_models()

    def _initialize_models(self):
        self.health_model = RandomForestRegressor()
        self.risk_model = RandomForestRegressor()
        self.scaler = StandardScaler()

    def analyze_financial_health(self, user_data):
        features = self._prepare_features(user_data)
        health_score = self._predict_health_score(features)
        risk_level = self._predict_risk_level(features)
        savings_projection = self._project_savings(user_data)
        
        return {
            'health_score': health_score,
            'risk_level': risk_level,
            'savings_projection': savings_projection,
            'analysis': self._generate_ai_analysis(user_data, health_score, risk_level, savings_projection)
        }

    def _prepare_features(self, user_data):
        features = []
        if user_data:
            features.extend([
                user_data.get('income', 0),
                user_data.get('savings', 0),
                user_data.get('debt', 0)
            ])
        return np.array(features).reshape(1, -1)

    def _predict_health_score(self, features):
        # Placeholder for actual prediction logic
        return 75.0

    def _predict_risk_level(self, features):
        # Placeholder for actual prediction logic
        return "Moderate"

    def _project_savings(self, user_data):
        # Placeholder for actual projection logic
        return 10000.0

    def _generate_ai_analysis(self, user_data, health_score, risk_level, savings_projection):
        # Placeholder for actual AI analysis
        return "Your financial health is in good shape. Consider increasing your savings rate."

    def simulate_expense_reduction(self, user_data, expense_type, reduction_percentage):
        # Placeholder for expense reduction simulation
        return {
            'scenario': 'expense_reduction',
            'type': expense_type,
            'reduction': reduction_percentage,
            'impact': 'positive',
            'analysis': f'Reducing {expense_type} by {reduction_percentage}% would improve your financial health.'
        }

    def simulate_income_increase(self, user_data, increase_percentage):
        # Placeholder for income increase simulation
        return {
            'scenario': 'income_increase',
            'increase': increase_percentage,
            'impact': 'positive',
            'analysis': f'Increasing income by {increase_percentage}% would significantly improve your financial position.'
        }

    def simulate_loan(self, user_data, amount, term, interest_rate, purpose):
        # Placeholder for loan simulation
        return {
            'scenario': 'loan',
            'amount': amount,
            'term': term,
            'interest_rate': interest_rate,
            'purpose': purpose,
            'impact': 'neutral',
            'analysis': f'A {purpose} loan of ${amount} over {term} months at {interest_rate}% interest.'
        }

    def simulate_investment(self, user_data, amount, risk_level, duration, strategy):
        # Placeholder for investment simulation
        return {
            'scenario': 'investment',
            'amount': amount,
            'risk_level': risk_level,
            'duration': duration,
            'strategy': strategy,
            'impact': 'positive',
            'analysis': f'Investing ${amount} in a {risk_level} {strategy} strategy over {duration}.'
        }

    def generate_recommendations(self, user_data, predictions, simulations):
        # Placeholder for recommendation generation
        return {
            'savings': ['Increase your emergency fund', 'Consider a high-yield savings account'],
            'investments': ['Diversify your portfolio', 'Consider index funds'],
            'debt': ['Pay off high-interest debt first', 'Consider debt consolidation']
        } 