requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.6.1
prophet==1.1.5
plotly==5.18.0
python-dotenv==1.0.0
openai==1.7.0
langchain==0.1.4
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 